+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: vicalvez <vicalvez@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 15:49:14 by vicalvez          #+#    #+#             */
-/*   Updated: 2023/11/15 14:24:50 by vicalvez         ###   ########.fr       */
+/*   Updated: 2023/11/29 10:25:27 by vicalvez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	i = 0;
 	if (!s)
 		return (NULL);
-	if ((int) start >= ft_strlen((char *)s))
+	if (start >= ft_strlen((char *)s))
 		len = 0;
 	if (ft_strlen((char *)s) - start < len)
 		len = ft_strlen((char *)s) - start;
